move update_similar_snippets to parent class + add to client_ui_sqlite
