+++ ui/backend/client_ui_sqlite.py
@@ -1,5 +1,9 @@
 from credentialdigger import SqliteClient
 from credentialdigger.client import Discovery
+from credentialdigger.snippet_similarity import (
+                 build_embedding_model,
+                 compute_similarity,
+                 compute_snippet_embedding)                                
 
 from .client_ui import UiClient
 
@@ -165,3 +169,32 @@ class SqliteUiClient(UiClient, SqliteClient):
                 " FROM discoveries WHERE repo_url=?"
                 " GROUP BY file_name"
             ))
+
+
+    def update_similar_snippets(self,
+                                target_snippet,
+                                state,
+                                repo_url,
+                                file_name=None,
+                                threshold=0.95):
+        discoveries = self.get_discoveries(repo_url,
+                                           file_name,
+                                           state_filter=state)[0]
+        model = build_embedding_model()
+        target_snippet_embedding = compute_snippet_embedding(
+                                 target_snippet,
+                                 model)
+        n_updated_snippets = 0
+        for d in discoveries:
+            if d['state'] == 'new':
+                snippet_embedding = compute_snippet_embedding(
+                                  [d['snippet']],
+                                  model)
+                similarity = compute_similarity(
+                           target_snippet_embedding,
+                           snippet_embedding)
+            if similarity > threshold:
+                n_updated_snippets += 1
+                self.update_discovery(d['id'],state)
+         return n_updated_snippets
+
